% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataFred.R
\name{getDataFred}
\alias{getDataFred}
\title{Load a Matrix}
\usage{
getDataFred(seriesId, apiKey, .freq = NULL, .returnVintages = FALSE)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
Load a Matrix
}
